under_water-image-enhancement-master/requirements.txt
psycopg2==2.7.2
appdirs==1.4.4
appJar==0.94.0
asgiref==3.2.10
bcrypt==3.2.0
beautifulsoup4==4.9.1
brute==0.0.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
comtypes==1.1.7
cryptography==3.2.1
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1
docutils==0.16
easyAI==1.0.0.4
filelock==3.0.12
future==0.18.2
gTTS==2.1.1
gTTS-token==1.1.3
gunicorn==20.0.4
idna==2.8
imageio==2.9.0
joblib==0.15.1
json5==0.9.5
kivy-deps.angle==0.1.10
kivy-deps.glew==0.1.12
kivy-deps.gstreamer==0.1.18
kivy-deps.sdl2==0.1.23
Kivy-examples==1.11.1
kiwisolver==1.2.0
matlab==0.1
matplotlib @ file:///C:/Users/shamil/Downloads/matplotlib-3.3.0rc1-cp38-cp38-win32.whl
mysqlclient @ file:///C:/Users/shamil/Downloads/mysqlclient-1.4.6-cp38-cp38-win32.whl
natsort==7.0.1
networkx==2.5
nltk==3.5
numpy==1.19.0
opencv-python==4.2.0.34
pandas==1.0.5
paramiko==2.7.2
pickle-mixin==1.0.2
Pillow==7.1.2
ply==3.11
pocketsphinx @ file:///C:/Users/shamil/Downloads/pocketsphinx-0.1.15-cp38-cp38-win32.whl
pyasn1==0.4.8
PyAudio @ file:///C:/Users/shamil/Downloads/PyAudio-0.2.11-cp38-cp38-win32.whl
pycparser==2.20
pycryptodome==3.9.8
pycryptodomex==3.9.8
Pygments==2.7.1
PyNaCl==1.4.0
pyparsing==2.4.7
pypiwin32==223
pyreadline==2.1
pysmi==0.3.4
pysnmp==4.4.6
python-dateutil==2.8.1
pyttsx==1.1
pyttsx3==2.88
pytz==2020.1
PyWavelets==1.1.1
pywin32==228
PyYAML==5.3.1
regex==2020.6.8
requests==2.21.0
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.5.0
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
sqlparse==0.3.1
threadpoolctl==2.1.0
tifffile==2020.11.18
tqdm==4.46.1
unrar==0.4
urllib3==1.24.3
virtualenv==20.0.35
whitenoise==5.2.0
xlwt==1.3.0
